==> notebooks/6_recommender_test.ipynb <==
## RECOMMENDER TEST

1. Load the dataset that contains our full dataset with clusters, including the columns that we need to display the songs in the app (album_cover_url, name, title, release_date, popularity, duration_ms, explicit ...)
2. Use Spotify API to get the track details for the user's selected song, using the sp.search() method (search endpoint). Use results to display to the user the list of songs that match the user's query.
3. Build the dataframe with the same song features / columns that we used for the clustering. Make sure the columns are in the same order as the columns in the clustering dataset. Make sure the transformations you apply to calculate genre columns are consistent with the ones used for the clustering.
4. Scale the data with the same scaler used for the clustering.
5. Predict the cluster index for the user's selected song.
6. Filter the dataset to get the songs from the same cluster.
7. Get the top 10 most popular songs from the same cluster.
8. Display the songs to the user.

from dotenv import load_dotenv
import os
import time
import spotipy
from spotipy.oauth2 import SpotifyClientCredentials
import pandas as pd
import numpy as np
from sklearn.preprocessing import StandardScaler
from sklearn.cluster import KMeans
from sklearn.metrics import silhouette_score
import pickle
from spotipy.oauth2 import SpotifyOAuth


# Load environment variables from .env file

# Load environment variables from .env file
load_dotenv()

# Now you can access the environment variables
client_id = os.getenv('SPOTIFY_CLIENT_ID')
client_secret = os.getenv('SPOTIFY_CLIENT_SECRET')

spotify = spotipy.Spotify(client_credentials_manager=SpotifyClientCredentials(client_id, client_secret))



## Load the dataset

#load the dataset
spotify_df = pd.read_csv('../data/7_clustered_dataset.csv')
spotify_df.head()



spotify_df.columns

## Spotify API  user query

def search_spotify_tracks(query, limit=10):
    """
    Search for tracks on Spotify and return formatted results
    
    Parameters:
    query (str): Search query for the track
    limit (int): Maximum number of results to return
    
    Returns:
    list: List of dictionaries containing track information
    """
    results = spotify.search(q=query, type='track', limit=limit)
    
    tracks = []
    for track in results['tracks']['items']:
        track_info = {
            'title': track['name'],
            'artist': track['artists'][0]['name'],
            'artist_id': track['artists'][0]['id'],
            'duration_ms': track['duration_ms'],
            'explicit': track['explicit'],
            'album': track['album']['name'],
            'album_cover': track['album']['images'][0]['url'] if track['album']['images'] else None,
            'preview_url': track['preview_url'],
            'popularity': track['popularity'],
            'id': track['id'],
            'release_year': track['album']['release_date'][:4]
        }
        tracks.append(track_info)
    
    return tracks


# Example usage: user types a song, then we search features by track id

query = "Born to die" 
search_results = search_spotify_tracks(query)
search_results


## Dataframe with the same song features

Bulid the dataframe with the same song features / columns that we used for the clustering. 

Columns must be in the same order as the columns in the clustering dataset.

Transformations to calculate genre columns must beconsistent with the ones used for the clustering.

columns_to_keep = ['popularity','duration_ms', 'explicit', 'pop','old school hip hop','rap','rock' ,'hard rock', 'release_year']

#user selects a song i.e. "Born to die" index 0
user_selection_index = 0
user_selection_track_data= search_results[user_selection_index]

#get genre from artist id
results=spotify.artist(user_selection_track_data['artist_id']) 

genres=results['genres']

print(genres)
print(user_selection_track_data)

#create a function to get necessary columns, and loop to get the genre
def get_necessary_columns(search_results):
    user_selection_index = 0
    user_selection_track_data= search_results[user_selection_index]

    #get genre from artist id
    results=spotify.artist(user_selection_track_data['artist_id']) 

    genres=results['genres']


    return {
        'popularity': user_selection_track_data['popularity'],
        'duration_ms': user_selection_track_data['duration_ms'],
        'explicit': user_selection_track_data['explicit'],
        'pop': 1 if 'pop' in genres else 0,
        'old school hip hop': 1 if 'old school hip hop' in genres else 0,
        'rap': 1 if 'rap' in genres else 0,
        'rock': 1 if 'rock' in genres else 0,
        'hard rock': 1 if 'hard rock' in genres else 0,
        'release_year': user_selection_track_data['release_year']
    }


# Dataframe with the same song features / columns that we used for the clustering.


selection_data = {
    'popularity': user_selection_track_data['popularity'],
    'duration_ms': user_selection_track_data['duration_ms'],
    'explicit': user_selection_track_data['explicit'],
    'pop': 1 if 'pop' in genres else 0,
    'old school hip hop': 1 if 'old school hip hop' in genres else 0,
    'rap': 1 if 'rap' in genres else 0,
    'rock': 1 if 'rock' in genres else 0,
    'hard rock': 1 if 'hard rock' in genres else 0,
    'release_year': user_selection_track_data['release_year']
}

selection_data_df = pd.DataFrame([selection_data])
selection_data_df


# Scale the data

# Load the scaler
scaler = pickle.load(open('../scaler/standard_scaler.pkl', 'rb'))

# Scale the data
selection_data_scaled = scaler.transform(selection_data_df)

selection_data_scaled_df = pd.DataFrame(selection_data_scaled, columns=selection_data_df.columns)

selection_data_scaled_df

# Using the pre-trained model

Getting the cluster index for the user's selected song and filtering top 10 songs in the same cluster

# Load model and predict
model = pickle.load(open('../models/kmeans_model_11.pkl', 'rb'))

cluster_prediction = model.predict(selection_data_scaled_df)
#cluster_prediction[0]

print(f'Cluster label: {cluster_prediction[0]}')

# Load the clustered dataset and filter for the cluster
clustered_df = pd.read_csv('../data/7_clustered_dataset.csv')
clustered_df_filtered = clustered_df[clustered_df['cluster'] == cluster_prediction[0]]

# Get the top 10 most popular songs from the same cluster
top_10_songs = clustered_df_filtered.sort_values(by='popularity', ascending=False).sample(10)
top_10_songs.head(10)
